restructure & add crud
