+++ src/main/java/usmb/info803/profile_search/entreprise/EntrepriseRepository.java
@@ -1,4 +1,4 @@
-package usmb.info803.profile_search.infrastructure.persistance;
+package usmb.info803.profile_search.entreprise;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
